refactor: split shared/utils.py into focused modules
Extract locally-defined functions into four new internal modules:
- _config.py (77 lines): Config dataclass and lazy singleton
- _env.py (103 lines): environment-based configuration accessors
- _repo.py (25 lines): repository URL helpers
- _diff.py (16 lines): diff display utility

utils.py (101 lines) remains as the public re-export facade.
