+++ src/quodeq/shared/_repo.py
@@ -1,0 +1,25 @@
+"""Repository URL helpers."""
+from __future__ import annotations
+
+from pathlib import Path
+
+
+def is_repo_url(repo_input: str) -> bool:
+    """Return True if the input looks like a remote repository URL.
+
+    Raises ValueError for cleartext ``http://`` URLs to enforce encrypted
+    transport for credential safety.
+    """
+    if repo_input.startswith("http://"):
+        raise ValueError(
+            "Cleartext HTTP repository URLs are rejected to protect credentials. "
+            "Use https:// or git@ instead."
+        )
+    return repo_input.startswith(("https://", "git@"))
+
+
+def project_name_from_repo(repo: str) -> str:
+    """Extract a human-readable project name from a repo path or URL."""
+    if is_repo_url(repo):
+        return repo.split("/")[-1].replace(".git", "")
+    return Path(repo).name
